backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3
solders==0.20.0
solana>=0.30.2